Extract GroupEdge from model_soup
