+++ grouper/fe/template_util.py
@@ -101,8 +101,8 @@ def get_template_env(package="grouper.fe", deployment_name="",
                      extra_filters=None, extra_globals=None):
     # TODO(herb): get around circular depdendencies; long term remove call to
     # send_async_email() from grouper.models
-    from grouper.model_soup import GROUP_EDGE_ROLES
     from grouper.models.base.constants import OBJ_TYPES_IDX
+    from grouper.models.group_edge import GROUP_EDGE_ROLES
 
     filters = {
         "print_date": print_date,
