+++ tests/test_grouper_ctl_group.py
@@ -5,8 +5,9 @@ from mock import patch
 
 from ctl_util import call_main
 from fixtures import standard_graph, graph, users, groups, session, permissions  # noqa
-from grouper.model_soup import Group, GROUP_EDGE_ROLES
+from grouper.model_soup import Group
 from grouper.models.audit_log import AuditLog
+from grouper.models.group_edge import GROUP_EDGE_ROLES
 
 
 @patch('grouper.ctl.group.make_session')
